This PR refactors and shares oss-fuzz specific test harness code with the afl fuzzer harness. ChangeLog updated.
