+++ test/tools/CMakeLists.txt
@@ -1,4 +1,8 @@
-add_executable(solfuzzer fuzzer.cpp)
+if (OSSFUZZ)
+    add_subdirectory(ossfuzz)
+endif()
+
+add_executable(solfuzzer afl_fuzzer.cpp fuzzer_common.cpp)
 target_link_libraries(solfuzzer PRIVATE libsolc evmasm ${Boost_PROGRAM_OPTIONS_LIBRARIES} ${Boost_SYSTEM_LIBRARIES})
 
 add_executable(yulopti yulopti.cpp)
